refactor: move the log form into it's own component
