+++ src/index.ts
@@ -1,2 +1,3 @@
 import './components/dominion-logger'
 import './components/log-list'
+import './components/log-form'
